Started moving layout header
