+++ src/bp/ui-shared/src/LayoutHeader/style.scss
