+++ src/bp/ui-studio/src/web/components/Layout/BottomPanel/style.scss.d.ts
